refactor(util): lift the plot type label lookup out of the model
The pre-activation instruction in the root component needed it, and a
component must not import the model.
